refactor: split NodeLink component into focused sub-components - Split monolithic NodeLink.tsx into separate, focused components:   - NodeLink: Main component orchestrating the link behavior   - NodeLinkIcon: Icon display logic   - NodeThumbnail: Thumbnail rendering with image preview support   - ThumbnailBox: Container for directory thumbnails   - ThumbnailFile: File-specific thumbnail rendering   - ThumbnailLabel: Label for directory item counts   - ThumbnailSheets: Stacked sheet visual effect for directories
- Update imports across DirectoryView components
- Convert NodeLink and Table to named exports

This refactoring improves code organization, testability, and reusability
without changing functionality.
